+++ update-build-tools-deps.sh
@@ -5,5 +5,7 @@ BASEDIR=$(dirname "$0")
 node $BASEDIR/merge-build-tools-deps.js
 node_modules/.bin/yarn --ignore-engines
 rc=$?
-mv package.json.bak package.json
+if [[ "$KEEP_MODIFIED_PACKAGE" != 'true' ]]; then
+    mv package.json.bak package.json
+fi
 exit $rc
